Split the GameLogicAPI.js into GameLogic.js and GameAPI.js to reflect their appropriate functionalities as well as reduce the amount of code in one file. Also added create_game_with_2_players fetch function in GameAPI.js and respetcive API endpoint in a_game/views.py. Will merge the invite to game logic from the front_oauth branch to fit and test the functionality..
