+++ _main_project/a_game/urls.py
@@ -8,4 +8,5 @@ urlpatterns = [
 	path("move_paddle/<str:game_id>/", views.move_paddle, name="move_paddle"),
 	path("end_game/<str:game_id>/", views.end_game_session, name="end_game"),
 	path("quit_game/", views.quit_game_session, name="quit_game"),
+	path("create_game_with_2_players/", views.create_game_with_2_players, name="create_game_with_2_players"),
 ]
